requirements.txt
supabase==2.0.3
python-dotenv==1.0.0
pandas==2.1.4
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0 